% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMOE}
\alias{getMOE}
\title{Calculate Modulus of elasticity from stress-strain curve}
\usage{
getMOE(stress, strain, plot = TRUE, robustness = 0.33, cleanup_width = 5)
}
\arguments{
\item{stress}{numeric, vector of values representing stress calculated from
function `stress_strain()`.}

\item{strain}{numeric, vector of values representing strain calculated from
function `stress_strain()`.}

\item{plot}{boolean, if `TRUE` plots a stress-strain diagram with fitted
line on linear portion representing modulus of elasticity}

\item{robustness, }{numeric between 0 and 1 representing the minimum proportion
of stress/strain readings to include in the linear fit. A value of 0.33 
represents that of values left of the peak, >=33% will be included in the model
fit.}

\item{cleanup_width, }{numeric, aids with gap filling missing data. When either 
stress or strain information is missing, fill in the value using the median
from the `cleanup_width` surrounding values.}
}
\description{
Function to estimate modulus of elasticity from the linear portion of a 
stress-strain curve. Function looks to 75% of the values to the left of 
peak stress, and identifies the linear portion through iterative linear
modeling until the best fit is found
}
\examples{
library(readr)
library(moeR)

#load data
tree = read_csv(system.file('extdata', 'tree_1.csv', package = 'moeR'))

#calculate turning moment
tree$M_kNm = tree$F_kN* tree$strap_ht_m

#calculate stress-strain diagram
ss = stress_strain(moment_kNm = tree$M_kNm,
  tilt_deg = tree$tilt,
  ht_m = 1.3,
  diam_cm = 20.2,
  plot = FALSE)
  
 # calculate modulus of elasticity
 moe = getMOE(ss$stress, ss$strain, plot=TRUE)
 print(moe)
}
