% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stress_strain}
\alias{stress_strain}
\title{Create data for stress-strain curve from tree winching data}
\usage{
stress_strain(moment_kNm, tilt_deg, ht_m, diam_cm, plot = TRUE)
}
\arguments{
\item{moment_kNm}{numeric, vector of values representing turning moment in
kNm. Usually calcualted as applied force (kN) from tensiometer times height
of strap attachment (m)}

\item{tilt_deg}{numeric, vector of values representing tilt measurements
from an angle gauge. Angles should be positive and  from vertical}

\item{ht_m}{numeric, height of tilt measurement}

\item{diam_cm}{numeric, diameter of tree at point of tilt measurement,
usually at the point of strap attachment}

\item{plot}{boolean, if `TRUE` plots a stress-strain diagram}
}
\description{
Function to calculate stress (xx) and strain (xx) information from tree winching
data as a pre-requisite for calculate modulus of elasticity.
}
\examples{
library(readr)
library(moeR)

#load data
tree = read_csv(system.file('extdata', 'tree_1.csv', package = 'moeR'))

#calculate turning moment
tree$M_kNm = tree$F_kN* tree$strap_ht_m

#calculate stress-strain diagram
ss = stress_strain(moment_kNm = tree$M_kNm,
  tilt_deg = tree$tilt,
  ht_m = 1.3,
  diam_cm = 20.2,
  plot = TRUE)
#drop na vals
ss = ss[!is.na(ss$stress),]

head(ss)
}
